refactor: rename state files to lowercase
