+++ src/screens/Departures/state/nearby-places-state.ts
